Moved hooks to root of the project
