+++ app/conversations/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import clsx from "clsx";
-import useConversation from "../hooks/useConversation";
+import useConversation from "../../hooks/useConversation";
 import EmptyState from "@/components/EmptyState";
 
 /**
