+++ components/sidebar/MobileFooter.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import useConversation from "@/app/hooks/useConversation";
-import useRoutes from "@/app/hooks/useRoutes";
+import useConversation from "@/hooks/useConversation";
+import useRoutes from "@/hooks/useRoutes";
 import MobileItem from "./MobileItem";
 
 /**
